Separate Controllers
done , working
